+++ examples/minimal-fastify-react/client/vite.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   plugins: [react()],
   server: {
     proxy: {
-      "/api": "http://localhost:3001",
+      "/api": "http://localhost:3000",
     },
   },
 });
