+++ examples/minimal-fastify-react/server/src/index.ts
@@ -20,7 +20,7 @@ fastify
 
 const start = async () => {
   try {
-    await fastify.listen({ port: 3001 });
+    await fastify.listen({ port: 3000 });
   } catch (err) {
     fastify.log.error(err);
     process.exit(1);
